Consolidate variables and expressions
There's really no point in having two separate files for these types,
since they depend on each other, and they constantly poke each others
internals.

This simplifies the build, and since there's a lot of inlining going on,
it also makes it easier for the compiler to optimize call sites.

+++ src/emeus-types-private.h
@@ -35,7 +35,7 @@ typedef enum {
 typedef struct {
   int ref_count;
 
-  int id_;
+  unsigned long id_;
 
   VariableType type;
